+++ release_notes.md
@@ -22,6 +22,7 @@ DFTools Changelog
 * Use the data size property when computing DAG edges weight in DAG transformation;
 * Adding vertex end reference to dag init vertices; 
 * Setting the INIT_SIZE property of a DAGInitVertex to the same value of a SDFInitVertex;
+* Copy all properties of an SDFVertex to a DAGVertex;
 
 ### Bug fix